refactor: decompose core/config/models.py into package
Split 39 Pydantic models from 794-line monolith into 8 focused submodules with full backward compatibility.
