--- openpaw/core/config/models.py
@@ -1,794 +1,0 @@
-"""Pydantic configuration models for OpenPaw.
-
-This module defines all configuration dataclasses used throughout OpenPaw.
-For loading and merging logic, see loader.py.
-"""
-
-from pathlib import Path
-from typing import Any, Literal
-
-from apscheduler.triggers.cron import CronTrigger
-from pydantic import BaseModel, Field, field_validator, model_validator
-
-from openpaw.core.paths import DOWNLOADS_DIR, SCREENSHOTS_DIR
-
-
-class QueueConfig(BaseModel):
-    """Configuration for the command queue system."""
-
-    mode: str = Field(default="collect", description="Default queue mode: steer, followup, collect")
-    debounce_ms: int = Field(default=1000, description="Debounce delay in milliseconds")
-    cap: int = Field(default=20, description="Max queued messages per session")
-    drop_policy: str = Field(default="summarize", description="Overflow policy: old, new, summarize")
-
-
-class LaneConfig(BaseModel):
-    """Configuration for queue lanes."""
-
-    main_concurrency: int = Field(default=4, description="Max concurrent runs in main lane")
-    subagent_concurrency: int = Field(default=8, description="Max concurrent runs in subagent lane")
-    cron_concurrency: int = Field(default=2, description="Max concurrent runs in cron lane")
-
-
-
-
-class AgentConfig(BaseModel):
-    """Configuration for agent defaults."""
-
-    model: str = Field(default="anthropic:claude-sonnet-4-20250514", description="Default model identifier")
-    api_key: str | None = Field(default=None, description="API key for the model provider")
-    max_turns: int = Field(default=50, description="Max agent turns per run")
-    temperature: float = Field(default=0.7, description="Model temperature")
-
-
-
-
-class ProviderDefinition(BaseModel):
-    """Named provider in the global catalog.
-
-    Defines connection details (type, api_key, base_url, region) that can be
-    referenced by name from workspace model configurations.
-    """
-
-    type: str | None = Field(default=None, description="LangChain provider type. Defaults to catalog key name.")
-    api_key: str | None = Field(default=None, description="API key for the provider")
-    base_url: str | None = Field(default=None, description="Custom API endpoint URL")
-    region: str | None = Field(default=None, description="AWS region for Bedrock models")
-
-    model_config = {"extra": "allow"}
-
-
-class WorkspaceModelConfig(BaseModel):
-    """LLM configuration for a workspace agent."""
-
-    provider: str | None = Field(default=None, description="Model provider (anthropic, openai, bedrock_converse, etc.)")
-    model: str | None = Field(default=None, description="Model identifier")
-    api_key: str | None = Field(default=None, description="API key for the model provider")
-    temperature: float | None = Field(default=None, description="Model temperature")
-    max_turns: int | None = Field(default=None, description="Max agent turns per run")
-    region: str | None = Field(default=None, description="AWS region for Bedrock models (e.g., us-east-1)")
-    max_output_tokens: int | None = Field(
-        default=None,
-        description="Maximum output tokens per LLM call (None = provider default). Applied to all agent types.",
-    )
-    max_retries: int | None = Field(
-        default=3,
-        description=(
-            "Max retry attempts on transient API errors (504, 429, connection errors). "
-            "None or 0 disables retries. Bedrock is excluded (boto3 manages its own retries)."
-        ),
-    )
-
-    @model_validator(mode="before")
-    @classmethod
-    def split_combined_model_string(cls, data: Any) -> Any:
-        """Auto-split 'provider:model_id' into separate fields.
-
-        Supports combined format like "anthropic:claude-sonnet-4-20250514"
-        when provider is not explicitly set. Only splits on the first colon
-        to handle Bedrock IDs like "us.anthropic.claude-haiku:v1:0".
-        """
-        if not isinstance(data, dict):
-            return data
-        model_val = data.get("model")
-        provider_val = data.get("provider")
-        if model_val and provider_val is None and isinstance(model_val, str) and ":" in model_val:
-            provider, _, model_id = model_val.partition(":")
-            data["provider"] = provider
-            data["model"] = model_id
-        return data
-
-    model_config = {"extra": "allow"}
-
-
-class ChannelLogConfig(BaseModel):
-    """Configuration for persistent channel message logging."""
-
-    enabled: bool = Field(default=True, description="Enable channel message logging to JSONL")
-    retention_days: int = Field(default=30, description="Days before logs are archived")
-
-    @field_validator("retention_days")
-    @classmethod
-    def validate_retention(cls, v: int) -> int:
-        """Validate retention_days is at least 1."""
-        if v < 1:
-            raise ValueError("retention_days must be at least 1")
-        return v
-
-
-class WorkspaceChannelConfig(BaseModel):
-    """Channel binding configuration for a workspace agent."""
-
-    name: str | None = Field(
-        default=None,
-        description="Unique channel name (defaults to type). Required when multiple channels share a type.",
-    )
-    type: str | None = Field(default=None, description="Channel type (telegram, discord, etc.)")
-    token: str | None = Field(default=None, description="Channel bot token")
-    allowed_users: list[int] = Field(default_factory=list, description="Allowed user IDs")
-    allowed_groups: list[int] = Field(default_factory=list, description="Allowed group IDs")
-    allow_all: bool = Field(default=False, description="Allow all users (insecure, use with caution)")
-    mention_required: bool = Field(
-        default=False,
-        description="Only respond in group chats when the bot is @mentioned. DMs always pass through.",
-    )
-    triggers: list[str] = Field(
-        default_factory=list,
-        description=(
-            "Trigger keywords for group chat filtering. If set, messages must contain"
-            " at least one keyword (case-insensitive). Uses OR logic with mention_required."
-        ),
-    )
-    user_aliases: dict[int, str] = Field(
-        default_factory=dict,
-        description="Map user IDs to display names for message attribution",
-    )
-    context_messages: int = Field(
-        default=25,
-        description=(
-            "Number of recent channel messages to fetch as context on trigger"
-            " (0 = disabled, max 100)"
-        ),
-    )
-    channel_log: ChannelLogConfig = Field(
-        default_factory=ChannelLogConfig,
-        description="Persistent channel message logging configuration",
-    )
-
-    @field_validator("context_messages")
-    @classmethod
-    def validate_context_messages(cls, v: int) -> int:
-        """Validate context_messages is between 0 and 100."""
-        if v < 0 or v > 100:
-            raise ValueError("context_messages must be between 0 and 100")
-        return v
-
-    model_config = {"extra": "allow"}
-
-
-class WorkspaceQueueConfig(BaseModel):
-    """Queue configuration overrides for a workspace agent."""
-
-    mode: str | None = Field(default=None, description="Queue mode: steer, followup, collect")
-    debounce_ms: int | None = Field(default=None, description="Debounce delay in milliseconds")
-
-    model_config = {"extra": "allow"}
-
-
-class BuiltinItemConfig(BaseModel):
-    """Configuration for a single builtin capability."""
-
-    enabled: bool = Field(default=True, description="Whether this builtin is active")
-    config: dict[str, Any] = Field(default_factory=dict, description="Builtin-specific settings")
-
-    model_config = {"extra": "allow"}
-
-
-class CronBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the cron scheduling tool."""
-
-    max_tasks: int = Field(default=50, description="Maximum scheduled tasks per workspace")
-    min_interval_seconds: int = Field(
-        default=300, description="Minimum interval between recurring tasks (5 min default, can be lowered)"
-    )
-
-
-class CronManagerBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the persistent cron management builtin."""
-
-
-class AcknowledgeBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the acknowledge_event tool."""
-
-
-class SendFileBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the send_file tool."""
-
-    max_file_size: int = Field(
-        default=50 * 1024 * 1024,
-        description="Maximum file size in bytes (default 50MB for Telegram)"
-    )
-
-
-class DoclingBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the Docling document processor."""
-
-    max_file_size: int = Field(
-        default=50 * 1024 * 1024,
-        description="Maximum file size in bytes (default 50MB)"
-    )
-    ocr_backend: str = Field(
-        default="auto",
-        description="OCR backend: 'auto', 'mac', 'easyocr', 'tesseract', 'rapidocr'"
-    )
-    ocr_languages: list[str] = Field(
-        default_factory=lambda: ["en"],
-        description="OCR languages as ISO 639-1 codes (auto-mapped per backend)"
-    )
-    force_full_page_ocr: bool = Field(
-        default=True,
-        description="Force full-page OCR (recommended for scanned docs)"
-    )
-    document_timeout: float | None = Field(
-        default=None,
-        description="Per-document timeout in seconds (None = no limit)"
-    )
-    do_ocr: bool = Field(default=True, description="Enable OCR processing")
-    do_table_structure: bool = Field(default=True, description="Enable table structure detection")
-
-
-class BrowserBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the browser automation tool."""
-
-    headless: bool = Field(default=True, description="Run browser without GUI")
-    allowed_domains: list[str] = Field(default_factory=list, description="Domain allowlist (empty = allow all)")
-    blocked_domains: list[str] = Field(default_factory=list, description="Domain blocklist (takes precedence)")
-    timeout_seconds: int = Field(default=30, description="Default timeout for browser operations")
-    persist_cookies: bool = Field(default=False, description="Persist cookies across agent runs")
-    downloads_dir: str = Field(default=str(DOWNLOADS_DIR), description="Directory for downloaded files")
-    screenshots_dir: str = Field(default=str(SCREENSHOTS_DIR), description="Directory for screenshots")
-
-
-class SpawnBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the sub-agent spawn tool."""
-
-    max_concurrent: int = Field(default=8, description="Maximum simultaneous sub-agents")
-    default_progress_interval: int = Field(
-        default=0,
-        ge=0,
-        description="Default progress interval in minutes (0 = disabled)",
-    )
-
-
-class Md2pdfBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the md2pdf conversion tool."""
-
-    theme: str = Field(default="minimal", description="CSS theme: minimal, professional, or technical")
-    max_diagram_width: float = Field(default=6.5, description="Maximum Mermaid diagram width in inches")
-    self_heal: bool = Field(default=True, description="Enable AI self-healing for broken Mermaid diagrams")
-    self_heal_model: str = Field(default="gpt-4o-mini", description="Model for self-healing (e.g., gpt-4o-mini)")
-    max_heal_iterations: int = Field(default=3, description="Maximum repair attempts per broken diagram")
-
-
-class ChannelHistoryBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the channel history browsing tool."""
-
-    max_messages_per_request: int = Field(
-        default=100,
-        description="Hard cap on messages returned per request (default: 100)",
-    )
-    content_truncation: int = Field(
-        default=500,
-        description="Per-message content character truncation limit (default: 500)",
-    )
-
-
-class GptResearcherBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the GPT-Researcher integration tool."""
-
-    endpoint: str = Field(default="", description="WebSocket endpoint URL (e.g., wss://researcher.example.com/ws)")
-    upload_endpoint: str = Field(default="", description="REST upload endpoint URL (e.g., https://researcher.example.com/upload/)")
-    timeout_seconds: int = Field(default=300, description="Max seconds to wait for research completion")
-    default_report_type: str = Field(
-        default="research_report",
-        description="Default report type: research_report, outline_report, detailed_report, resource_report",
-    )
-    default_report_source: str = Field(default="web", description="Default report source: web or local")
-    default_tone: str = Field(default="Objective", description="Default writing tone for reports")
-
-
-class EmailBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the email integration tool."""
-
-    provider: str = Field(default="gmail", description="Email provider: gmail")
-    service_account_file: str | None = Field(
-        default=None,
-        description="Path to Google service account JSON (absolute or relative to workspace)",
-    )
-    delegated_user: str | None = Field(
-        default=None,
-        description="Email address to impersonate via domain-wide delegation",
-    )
-    allowed_recipients: list[str] = Field(
-        default_factory=list,
-        description="Recipient allowlist patterns (e.g., '*@company.com'). Empty = block all sends.",
-    )
-    max_recipients: int = Field(
-        default=10,
-        description="Maximum recipients per email (to + cc + bcc combined)",
-    )
-
-
-class FilePersistenceBuiltinConfig(BuiltinItemConfig):
-    """Configuration for the file persistence processor."""
-
-    max_file_size: int = Field(
-        default=50 * 1024 * 1024,
-        description="Maximum file size in bytes (default 50MB)",
-    )
-    clear_data_after_save: bool = Field(default=False, description="Free memory after saving")
-
-
-class BuiltinsConfig(BaseModel):
-    """Global builtins configuration.
-
-    Supports OpenClaw-style allow/deny lists with group prefixes (e.g., "group:voice").
-    Deny takes precedence over allow. Empty allow list means allow all available.
-    """
-
-    allow: list[str] = Field(
-        default_factory=list,
-        description="Allowed builtins/groups (empty = allow all available)",
-    )
-    deny: list[str] = Field(
-        default_factory=list,
-        description="Denied builtins/groups (takes precedence over allow)",
-    )
-
-    # Per-builtin configuration
-    brave_search: BuiltinItemConfig = Field(default_factory=BuiltinItemConfig)
-    whisper: BuiltinItemConfig = Field(default_factory=BuiltinItemConfig)
-    elevenlabs: BuiltinItemConfig = Field(default_factory=BuiltinItemConfig)
-    shell: BuiltinItemConfig = Field(default_factory=BuiltinItemConfig)
-    cron: CronBuiltinConfig = Field(default_factory=CronBuiltinConfig)
-    cron_manager: CronManagerBuiltinConfig = Field(default_factory=CronManagerBuiltinConfig)
-    acknowledge: AcknowledgeBuiltinConfig = Field(default_factory=AcknowledgeBuiltinConfig)
-    send_file: SendFileBuiltinConfig = Field(default_factory=SendFileBuiltinConfig)
-    docling: DoclingBuiltinConfig = Field(default_factory=DoclingBuiltinConfig)
-    browser: BrowserBuiltinConfig = Field(default_factory=BrowserBuiltinConfig)
-    spawn: SpawnBuiltinConfig = Field(default_factory=SpawnBuiltinConfig)
-    file_persistence: FilePersistenceBuiltinConfig = Field(default_factory=FilePersistenceBuiltinConfig)
-    md2pdf: Md2pdfBuiltinConfig = Field(default_factory=Md2pdfBuiltinConfig)
-    channel_history: ChannelHistoryBuiltinConfig = Field(default_factory=ChannelHistoryBuiltinConfig)
-    gpt_researcher: GptResearcherBuiltinConfig = Field(default_factory=GptResearcherBuiltinConfig)
-    email: EmailBuiltinConfig = Field(default_factory=EmailBuiltinConfig)
-
-    model_config = {"extra": "allow"}
-
-
-class WorkspaceBuiltinsConfig(BaseModel):
-    """Per-workspace builtins configuration (overrides global)."""
-
-    allow: list[str] = Field(
-        default_factory=list,
-        description="Additional allowed builtins for this workspace",
-    )
-    deny: list[str] = Field(
-        default_factory=list,
-        description="Builtins to disable for this workspace",
-    )
-
-    # Per-builtin overrides
-    brave_search: BuiltinItemConfig | None = None
-    whisper: BuiltinItemConfig | None = None
-    elevenlabs: BuiltinItemConfig | None = None
-    shell: BuiltinItemConfig | None = None
-    cron: CronBuiltinConfig | None = None
-    cron_manager: CronManagerBuiltinConfig | None = None
-    acknowledge: AcknowledgeBuiltinConfig | None = None
-    send_file: SendFileBuiltinConfig | None = None
-    docling: DoclingBuiltinConfig | None = None
-    browser: BrowserBuiltinConfig | None = None
-    spawn: SpawnBuiltinConfig | None = None
-    file_persistence: FilePersistenceBuiltinConfig | None = None
-    md2pdf: Md2pdfBuiltinConfig | None = None
-    channel_history: ChannelHistoryBuiltinConfig | None = None
-    gpt_researcher: GptResearcherBuiltinConfig | None = None
-    email: EmailBuiltinConfig | None = None
-
-    model_config = {"extra": "allow"}
-
-
-class WorkspaceToolsConfig(BaseModel):
-    """Allow/deny list for workspace tools loaded from tools/ directory."""
-
-    allow: list[str] = Field(default_factory=list, description="Allowed tool names (empty = allow all)")
-    deny: list[str] = Field(default_factory=list, description="Denied tool names")
-
-
-class HeartbeatConfig(BaseModel):
-    """Configuration for the heartbeat scheduler."""
-
-    enabled: bool = Field(default=False, description="Enable periodic heartbeat prompts")
-    interval_minutes: int = Field(default=30, description="Minutes between heartbeat checks")
-    active_hours: str | None = Field(
-        default=None,
-        description="Active hours window (e.g., '08:00-22:00'). None = always active",
-    )
-    suppress_ok: bool = Field(default=True, description="Suppress HEARTBEAT_OK responses from channel")
-    target_channel: str = Field(default="telegram", description="Channel name to route heartbeat responses")
-    target_id: int | None = Field(default=None, description="Target ID for output routing (preferred)")
-    target_chat_id: int | None = Field(default=None, description="Telegram chat ID (deprecated, use target_id)")
-    target_channel_id: int | None = Field(default=None, description="Discord channel ID (deprecated, use target_id)")
-    delivery: Literal["channel", "agent"] = Field(
-        default="channel",
-        description="Where to deliver results: channel (direct) or agent (queue injection)",
-    )
-    max_output_tokens: int | None = Field(
-        default=None,
-        description="Max output tokens for heartbeat runs (overrides workspace model default)",
-    )
-
-    @field_validator("delivery", mode="before")
-    @classmethod
-    def validate_delivery(cls, v: str) -> str:
-        """Validate delivery mode and reject removed 'both' option."""
-        if v == "both":
-            raise ValueError(
-                "delivery: 'both' has been removed. Use 'agent' to have the main agent "
-                "process heartbeat results (it can use acknowledge_event to suppress "
-                "routine output), or 'channel' for direct delivery."
-            )
-        return v
-
-
-class ToolApprovalConfig(BaseModel):
-    """Configuration for a single tool's approval requirements."""
-
-    require_approval: bool = Field(
-        default=True,
-        description="Whether this tool requires user approval before execution",
-    )
-    show_args: bool = Field(
-        default=True,
-        description="Whether to show tool arguments in the approval prompt",
-    )
-
-
-class ApprovalGatesConfig(BaseModel):
-    """Root configuration for the approval gates system."""
-
-    enabled: bool = Field(
-        default=False,
-        description="Whether approval gates are active for this workspace",
-    )
-    timeout_seconds: int = Field(
-        default=120,
-        description="Seconds to wait for user approval before applying default action",
-    )
-    default_action: str = Field(
-        default="deny",
-        description="Action when approval times out: 'deny' or 'approve'",
-    )
-    tools: dict[str, ToolApprovalConfig] = Field(
-        default_factory=dict,
-        description="Per-tool approval configuration (tool_name -> config)",
-    )
-
-
-class EmbeddingConfig(BaseModel):
-    """Configuration for the embedding provider."""
-
-    provider: str = Field(default="openai", description="Embedding provider (openai)")
-    model: str = Field(default="text-embedding-3-small", description="Embedding model name")
-    api_key: str | None = Field(default=None, description="API key for embedding provider")
-
-
-class VectorStoreConfig(BaseModel):
-    """Configuration for the vector store backend."""
-
-    provider: str = Field(
-        default="sqlite_vec", description="Vector store provider (sqlite_vec)"
-    )
-    dimensions: int = Field(default=1536, description="Embedding dimensions")
-
-
-class MemoryConfig(BaseModel):
-    """Configuration for conversation memory and vector search."""
-
-    enabled: bool = Field(default=False, description="Enable conversation vector search")
-    vector_store: VectorStoreConfig = Field(
-        default_factory=VectorStoreConfig, description="Vector store backend config"
-    )
-    embedding: EmbeddingConfig = Field(
-        default_factory=EmbeddingConfig, description="Embedding provider config"
-    )
-
-
-class ToolTimeoutsConfig(BaseModel):
-    """Configuration for per-tool-call timeouts."""
-
-    default_seconds: int = Field(
-        default=120,
-        description="Default timeout for tool calls in seconds",
-    )
-    overrides: dict[str, int] = Field(
-        default_factory=dict,
-        description="Per-tool timeout overrides (tool_name -> timeout_seconds)",
-    )
-
-
-class AutoCompactConfig(BaseModel):
-    """Configuration for automatic context compaction."""
-
-    enabled: bool = Field(default=False, description="Enable automatic compaction when context window fills")
-    trigger: float = Field(default=0.8, description="Context utilization fraction to trigger compaction (0.0-1.0)")
-    summary_model: str | None = Field(
-        default=None, description="Model for summary generation (null = use workspace model)"
-    )
-
-    @field_validator("trigger")
-    @classmethod
-    def validate_trigger(cls, v: float) -> float:
-        """Validate trigger is a fraction between 0.0 and 1.0."""
-        if not 0.0 <= v <= 1.0:
-            raise ValueError("trigger must be between 0.0 and 1.0")
-        return v
-
-
-class StatusReminderConfig(BaseModel):
-    """Configuration for status reminder middleware.
-
-    Controls when and how often the agent is reminded to call send_message()
-    during long silent tool-calling runs.
-    """
-
-    enabled: bool = Field(default=True, description="Enable silent operation detection")
-    threshold: int = Field(
-        default=5,
-        ge=1,
-        le=50,
-        description="Turns without send_message before reminding",
-    )
-    max_reminders: int = Field(
-        default=3,
-        ge=0,
-        le=20,
-        description="Max reminders per agent run",
-    )
-    cooldown_turns: int = Field(
-        default=1,
-        ge=0,
-        le=10,
-        description="Min turns between consecutive reminders",
-    )
-
-
-class LifecycleConfig(BaseModel):
-    """Configuration for lifecycle event notifications."""
-
-    notify_startup: bool = Field(default=False, description="Send notification when workspace starts")
-    notify_shutdown: bool = Field(default=True, description="Send notification when workspace stops")
-    notify_auto_compact: bool = Field(default=True, description="Send notification on auto-compact")
-    notify_session_ttl: bool = Field(default=True, description="Send notification on session TTL expiry")
-
-
-class CronOutputConfig(BaseModel):
-    """Output routing configuration for a cron job."""
-
-    channel: str = Field(description="Channel name (or type for backward compat)")
-    target_id: int | None = Field(default=None, description="Target ID for output routing (preferred)")
-    chat_id: int | None = Field(default=None, description="Telegram chat ID (deprecated, use target_id)")
-    guild_id: int | None = Field(default=None, description="Discord guild ID")
-    channel_id: int | None = Field(default=None, description="Discord channel ID (deprecated, use target_id)")
-    delivery: Literal["channel", "agent"] = Field(
-        default="channel",
-        description="Delivery mode: channel (direct) or agent (queue injection)",
-    )
-
-    @field_validator("delivery", mode="before")
-    @classmethod
-    def validate_delivery(cls, v: str) -> str:
-        """Validate delivery mode and reject removed 'both' option."""
-        if v == "both":
-            raise ValueError(
-                "delivery: 'both' has been removed. Use 'agent' to have the main agent "
-                "process cron results (it can use acknowledge_event to suppress routine "
-                "output), or 'channel' for direct delivery."
-            )
-        return v
-
-
-class CronDefinition(BaseModel):
-    """Definition of a single cron job from workspace crons/ directory."""
-
-    name: str = Field(description="Unique job identifier")
-    schedule: str = Field(description="Cron expression (e.g., '0 9 * * *')")
-    enabled: bool = Field(default=True, description="Whether the job is active")
-    prompt: str = Field(description="User prompt to inject when cron triggers")
-    output: CronOutputConfig = Field(description="Where to send the response")
-    max_output_tokens: int | None = Field(
-        default=None,
-        description="Max output tokens for this cron job (overrides workspace model default)",
-    )
-
-    @field_validator("schedule")
-    @classmethod
-    def validate_cron_expression(cls, v: str) -> str:
-        """Validate cron expression is parseable at config load time."""
-        try:
-            CronTrigger.from_crontab(v)
-        except (ValueError, KeyError) as e:
-            raise ValueError(f"Invalid cron expression '{v}': {e}") from e
-        return v
-
-
-class WorkspaceConfig(BaseModel):
-    """Configuration for a workspace agent (loaded from agent.yaml).
-
-    Supports shorthand ``model: "provider:model_id"`` in agent.yaml, which is
-    coerced to ``model: {provider: ..., model: ...}`` before validation.
-
-    Channels can be configured as a single ``channel:`` block (backward compat)
-    or as a ``channels:`` list for multi-channel workspaces. The model_validator
-    normalizes both forms into the ``channels`` list.
-    """
-
-    @model_validator(mode="before")
-    @classmethod
-    def normalize_channel_config(cls, data: Any) -> Any:
-        """Normalize channel/channels config and coerce model shorthand.
-
-        Handles:
-        - ``model: 'provider:model_id'`` shorthand → dict
-        - ``channel:`` (singular) → ``channels: [channel]``
-        - ``channels:`` (list) → used as-is
-        - Both present → ValueError
-        """
-        if not isinstance(data, dict):
-            return data
-
-        # Coerce model shorthand
-        if isinstance(data.get("model"), str):
-            model_str = data["model"]
-            if ":" in model_str:
-                p, _, m = model_str.partition(":")
-                data["model"] = {"provider": p, "model": m}
-            else:
-                data["model"] = {"model": model_str}
-
-        # Normalize channel → channels
-        has_channel = "channel" in data and data["channel"]
-        has_channels = "channels" in data and data["channels"]
-
-        if has_channel and has_channels:
-            raise ValueError(
-                "Cannot specify both 'channel' and 'channels' in workspace config. "
-                "Use 'channels' (list) for multi-channel or 'channel' (dict) for single."
-            )
-
-        if has_channel:
-            data["channels"] = [data.pop("channel")]
-        else:
-            data.pop("channel", None)
-
-        return data
-
-    timezone: str = Field(default="UTC", description="Workspace timezone (e.g., 'America/Los_Angeles')")
-    model: WorkspaceModelConfig = Field(default_factory=WorkspaceModelConfig, description="LLM configuration")
-    channels: list[WorkspaceChannelConfig] = Field(
-        default_factory=list,
-        description="Channel bindings (list). Use singular 'channel:' in YAML for backward compat.",
-    )
-    queue: WorkspaceQueueConfig = Field(default_factory=WorkspaceQueueConfig, description="Queue overrides")
-    builtins: WorkspaceBuiltinsConfig = Field(
-        default_factory=WorkspaceBuiltinsConfig,
-        description="Builtin capability overrides",
-    )
-    heartbeat: HeartbeatConfig | None = Field(default=None, description="Per-workspace heartbeat config")
-    workspace_tools: WorkspaceToolsConfig = Field(
-        default_factory=WorkspaceToolsConfig,
-        description="Allow/deny list for workspace tools",
-    )
-    approval_gates: ApprovalGatesConfig = Field(
-        default_factory=ApprovalGatesConfig,
-        description="Approval gates configuration",
-    )
-    tool_timeouts: ToolTimeoutsConfig = Field(
-        default_factory=ToolTimeoutsConfig,
-        description="Per-tool-call timeout configuration",
-    )
-    memory: MemoryConfig = Field(
-        default_factory=MemoryConfig,
-        description="Conversation memory and vector search configuration",
-    )
-    auto_compact: AutoCompactConfig = Field(
-        default_factory=AutoCompactConfig,
-        description="Auto-compact configuration",
-    )
-    session_ttl_minutes: int = Field(
-        default=180,
-        description="Auto-reset conversation after N minutes of inactivity (0 to disable)",
-    )
-    checkpoint_retention_days: int = Field(
-        default=7,
-        description="Prune orphaned checkpoint data older than N days at startup (0 to disable)",
-    )
-    lifecycle: LifecycleConfig = Field(
-        default_factory=LifecycleConfig,
-        description="Lifecycle notification configuration",
-    )
-    status_reminder: StatusReminderConfig = Field(
-        default_factory=StatusReminderConfig,
-        description="Status reminder middleware configuration",
-    )
-
-    @field_validator("session_ttl_minutes")
-    @classmethod
-    def validate_session_ttl_minutes(cls, v: int) -> int:
-        if v < 0:
-            raise ValueError("session_ttl_minutes must be >= 0")
-        return v
-
-    @field_validator("checkpoint_retention_days")
-    @classmethod
-    def validate_checkpoint_retention_days(cls, v: int) -> int:
-        if v < 0:
-            raise ValueError("checkpoint_retention_days must be >= 0")
-        return v
-
-    @field_validator("timezone")
-    @classmethod
-    def validate_timezone(cls, v: str) -> str:
-        """Validate timezone is a valid IANA timezone identifier."""
-        from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
-
-        try:
-            ZoneInfo(v)
-        except (ZoneInfoNotFoundError, KeyError):
-            raise ValueError(
-                f"Invalid timezone '{v}'. Use IANA timezone identifiers "
-                f"(e.g., 'America/Denver', 'Europe/London', 'UTC')."
-            )
-        return v
-
-    model_config = {"extra": "allow"}
-
-
-class LoggingConfig(BaseModel):
-    """Configuration for logging system."""
-
-    level: str = Field(default="INFO", description="Logging level (DEBUG, INFO, WARNING, ERROR, CRITICAL)")
-    directory: str = Field(default="logs", description="Directory for log files")
-    max_size_mb: int = Field(default=10, description="Maximum log file size in MB before rotation")
-    backup_count: int = Field(default=5, description="Number of backup log files to keep")
-    per_workspace: bool = Field(default=True, description="Create separate log files per workspace")
-
-
-class Config(BaseModel):
-    """Root configuration for OpenPaw."""
-
-    workspaces_path: Path = Field(default=Path("agent_workspaces"), description="Path to agent workspaces")
-    logging: LoggingConfig = Field(default_factory=LoggingConfig, description="Logging configuration")
-    providers: dict[str, ProviderDefinition] = Field(
-        default_factory=dict,
-        description="Named provider catalog for reusable connection details",
-    )
-    queue: QueueConfig = Field(default_factory=QueueConfig)
-    lanes: LaneConfig = Field(default_factory=LaneConfig)
-    agent: AgentConfig = Field(default_factory=AgentConfig)
-    builtins: BuiltinsConfig = Field(default_factory=BuiltinsConfig, description="Builtin capabilities config")
-    approval_gates: ApprovalGatesConfig = Field(
-        default_factory=ApprovalGatesConfig,
-        description="Default approval gates configuration",
-    )
-    tool_timeouts: ToolTimeoutsConfig = Field(
-        default_factory=ToolTimeoutsConfig,
-        description="Default tool timeout configuration",
-    )
-    team_profiles_path: str | None = Field(
-        default=None,
-        description="Path to system-level spawn profiles directory (available to all workspaces)",
-    )
-
-    model_config = {"extra": "allow"}
