+++ openpaw/core/config/models/__init__.py
@@ -1,0 +1,103 @@
+"""Pydantic configuration models for OpenPaw.
+
+This package defines all configuration dataclasses used throughout OpenPaw.
+For loading and merging logic, see loader.py.
+"""
+
+from openpaw.core.config.models.base import (
+    AgentConfig,
+    Config,
+    LaneConfig,
+    LoggingConfig,
+    ProviderDefinition,
+    QueueConfig,
+)
+from openpaw.core.config.models.builtin import (
+    AcknowledgeBuiltinConfig,
+    BrowserBuiltinConfig,
+    BuiltinItemConfig,
+    BuiltinsConfig,
+    ChannelHistoryBuiltinConfig,
+    CronBuiltinConfig,
+    CronManagerBuiltinConfig,
+    DoclingBuiltinConfig,
+    EmailBuiltinConfig,
+    FilePersistenceBuiltinConfig,
+    GptResearcherBuiltinConfig,
+    Md2pdfBuiltinConfig,
+    SendFileBuiltinConfig,
+    SpawnBuiltinConfig,
+    WorkspaceBuiltinsConfig,
+)
+from openpaw.core.config.models.channel import (
+    ChannelLogConfig,
+    WorkspaceChannelConfig,
+)
+from openpaw.core.config.models.cron import (
+    CronDefinition,
+    CronOutputConfig,
+    HeartbeatConfig,
+)
+from openpaw.core.config.models.lifecycle import (
+    LifecycleConfig,
+    StatusReminderConfig,
+)
+from openpaw.core.config.models.memory import (
+    AutoCompactConfig,
+    EmbeddingConfig,
+    MemoryConfig,
+    VectorStoreConfig,
+)
+from openpaw.core.config.models.security import (
+    ApprovalGatesConfig,
+    ToolApprovalConfig,
+    ToolTimeoutsConfig,
+)
+from openpaw.core.config.models.workspace import (
+    WorkspaceConfig,
+    WorkspaceModelConfig,
+    WorkspaceQueueConfig,
+    WorkspaceToolsConfig,
+)
+
+__all__ = [
+    "AgentConfig",
+    "ApprovalGatesConfig",
+    "AcknowledgeBuiltinConfig",
+    "AutoCompactConfig",
+    "BrowserBuiltinConfig",
+    "BuiltinItemConfig",
+    "BuiltinsConfig",
+    "ChannelHistoryBuiltinConfig",
+    "ChannelLogConfig",
+    "Config",
+    "CronBuiltinConfig",
+    "CronDefinition",
+    "CronManagerBuiltinConfig",
+    "CronOutputConfig",
+    "DoclingBuiltinConfig",
+    "EmailBuiltinConfig",
+    "EmbeddingConfig",
+    "FilePersistenceBuiltinConfig",
+    "GptResearcherBuiltinConfig",
+    "HeartbeatConfig",
+    "LaneConfig",
+    "LifecycleConfig",
+    "LoggingConfig",
+    "Md2pdfBuiltinConfig",
+    "MemoryConfig",
+    "ProviderDefinition",
+    "QueueConfig",
+    "SendFileBuiltinConfig",
+    "SpawnBuiltinConfig",
+    "StatusReminderConfig",
+    "ToolApprovalConfig",
+    "ToolTimeoutsConfig",
+    "VectorStoreConfig",
+    "WorkspaceBuiltinsConfig",
+    "WorkspaceChannelConfig",
+    "WorkspaceConfig",
+    "WorkspaceModelConfig",
+    "WorkspaceQueueConfig",
+    "WorkspaceToolsConfig",
+]
